Fix a signed-unsigned comparison
`LED_BYTES_PER_BANK` will be treated as an unsigned integer by the compiler, make
sure that the loop variable we compare against it is unsigned too. While `i`
here would never overflow, because `LED_BYTES_BANK` will always be smaller than
128, this does fix a compiler warning, and is the right thing to do anyways.

+++ KeyboardioScanner.cpp
@@ -152,7 +152,7 @@ void KeyboardioScanner::sendLEDData() {
 void KeyboardioScanner::sendLEDBank(byte bank) {
     uint8_t data[LED_BYTES_PER_BANK+1];
      data[0]  = TWI_CMD_LED_BASE+bank;
-	for(int i = 0 ; i < LED_BYTES_PER_BANK;i++) {
+	for(uint8_t i = 0 ; i < LED_BYTES_PER_BANK;i++) {
 		data[i+1] = ledData.bytes[bank][i];
 	}
     uint8_t result = twi_writeTo(addr,data,ELEMENTS(data) ,1,0);
